test(action-sheet): migrate trigger test
